[#931] Add `ConnectionLimitManager` + Metrics.getNumberOfConnections() returns int.
Add the `ConnectionLimitManager` and use it for MQTT and AMQP protocol
adapters.
AMQP: If no connection limit is configured, the AMQP protocol adapter auto-
configures a limit based on the amount of memory and the configured
maximum session window size. If a limit is configured and it is above
the recommended value, a warning is printed.
The MQTT adapter uses now ConnectionLimitManager.

+++ service-base/src/main/java/org/eclipse/hono/service/metric/MicrometerBasedMetrics.java
@@ -16,6 +16,7 @@ package org.eclipse.hono.service.metric;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Supplier;
 
@@ -65,7 +66,7 @@ public class MicrometerBasedMetrics implements Metrics {
 
     private final Map<String, AtomicLong> authenticatedConnections = new ConcurrentHashMap<>();
     private final AtomicLong unauthenticatedConnections;
-    private final AtomicLong totalCurrentConnections = new AtomicLong();
+    private final AtomicInteger totalCurrentConnections = new AtomicInteger();
 
     private LegacyMetrics legacyMetrics;
 
@@ -125,7 +126,7 @@ public class MicrometerBasedMetrics implements Metrics {
     }
 
     @Override
-    public long getNumberOfConnections() {
+    public int getNumberOfConnections() {
         return this.totalCurrentConnections.get();
     }
